tests/requirements-test.txt
# Testing Dependencies for Stocky Backend

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0

# HTTP testing
httpx>=0.24.0

# Test data factories
factory-boy>=3.3.0

# Time manipulation for testing
freezegun>=1.2.2

# Database testing utilities
pytest-alembic>=0.10.0

# Performance testing (optional)
pytest-benchmark>=4.0.0

# Test reporting and documentation
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Linting and code quality for tests
pytest-flake8>=1.1.1
pytest-mypy>=0.10.3
